1. 安裝npm; 2. 安裝 Laravel 的 ui 工具套件： composer require laravel/ui --dev 3. 建立含有使用者認證的bootstrap框架 php artisan ui bootstrap --auth 4. 安裝相依性套件並編譯 npm install && npm run dev npm run development 5.編輯 app/User.php，加入三種帳號角色名稱的常數字串，並將 role 這個欄位加入 fillable 中。 6. 編輯 database/migrations/2014_10_12_000000_create_users_table.php 這一個建立使用者資料表的 migration 設定檔，在其中加入一欄用於儲存帳號角色的 role 欄位。 7. 編輯 app/Http/Controllers/Auth/RegisterController.php，設定使用者註冊時的角色都是一般使用者。 8. 建立 Gate 權限規則。 9.自動生成三個 users: admin, manager, and user (方便我們做測試之用) 10.更新兩個views： players.index 和teams.index，讓他們針對admin 與 manager 呈現受控管的超連結(新增、修改、刪除) 11.讓每一個路由都必須經過 auth 的 middleware (即必須登入帳號才可以使用)。 12.增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用 13.不允許編輯表單給 user 權限使用，但允許 admin或manager使用 14.修改照片 15.背景顏色更新

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\BrandsController;
 use App\Http\Controllers\HeadphonesController;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -14,6 +15,8 @@ use App\Http\Controllers\HeadphonesController;
 |
 */
 //註解一下
+Route::middleware(['auth'])->group(function () {
+
 Route::get('/', function () {
     return redirect('brands');
 });
@@ -24,15 +27,15 @@ Route::get('brands/{id}',[BrandsController::class,'show'])->where('id','[0-9]+')
 // 修改單一球員表單
 Route::get('brands/{id}/edit',[BrandsController::class,'edit'])->where('id','[0-9]+')->name('brands.edit');
 // 刪除單一球隊及旗下球員資料
-Route::delete('brands/delete/{id}', [BrandsController::class, 'destroy'])->where('id', '[0-9]+')->name('brands.destroy');
+Route::delete('brands/delete/{id}', [BrandsController::class, 'destroy'])->where('id', '[0-9]+')->name('brands.destroy')->middleware('can:admin');
 // 新增球隊表單
-Route::get('brands/create', [BrandsController::class, 'create'])->name('brands.create');
+Route::get('brands/create', [BrandsController::class, 'create'])->name('brands.create')->middleware('can:admin');
 // 修改球隊表單
 Route::get('brands/{id}/edit', [BrandsController::class, 'edit'])->where('id', '[0-9]+')->name('brands.edit');
 // 修改球隊資料
 Route::patch('brands/update/{id}', [BrandsController::class, 'update'])->where('id', '[0-9]+')->name('brands.update');
 // 儲存新球隊資料
-Route::post('brands/store', [BrandsController::class, 'store'])->name('brands.store');
+Route::post('brands/store', [BrandsController::class, 'store'])->name('brands.store')->middleware('can:admin');
 
 Route::get('brands/classic', [BrandsController::class, 'classic'])->name('brands.classic');
 
@@ -46,15 +49,15 @@ Route::get('headphones/{id}',[HeadphonesController::class,'show'])->where('id','
 // 修改單一球隊表單
 Route::get('headphones/{id}/edit',[HeadphonesController::class,'edit'])->where('id','[0-9]+')->name('headphones.edit');
 //刪除單一球員資料
-Route::delete('headphones/delete/{id}', [HeadphonesController::class, 'destroy'])->where('id', '[0-9]+')->name('headphones.destroy');
+Route::delete('headphones/delete/{id}', [HeadphonesController::class, 'destroy'])->where('id', '[0-9]+')->name('headphones.destroy')->middleware('can:admin');
 // 新增球員表單
-Route::get('headphones/create', [HeadphonesController::class, 'create'])->name('headphones.create');
+Route::get('headphones/create', [HeadphonesController::class, 'create'])->name('headphones.create')->middleware('can:admin');
 // 修改球員表單
 Route::get('headphones/{id}/edit', [HeadphonesController::class, 'edit'])->where('id', '[0-9]+')->name('headphones.edit');
 // 修改球員資料
 Route::patch('headphones/update/{id}', [HeadphonesController::class, 'update'])->where('id', '[0-9]+')->name('headphones.update');
 // 儲存新球員資料
-Route::post('headphones/store', [HeadphonesController::class, 'store'])->where('id', '[0-9]+')->name('headphones.store');
+Route::post('headphones/store', [HeadphonesController::class, 'store'])->where('id', '[0-9]+')->name('headphones.store')->middleware('can:admin');
 //便宜耳機
 Route::get('headphones/cheappirce', [HeadphonesController::class, 'cheappirce'])->name('headphones.cheappirce');
 //奢華耳機
@@ -62,4 +65,10 @@ Route::get('headphones/eppirce', [HeadphonesController::class, 'eppirce'])->name
 //選定位置查詢耳機
 Route::get('headphones/genre', [HeadphonesController::class, 'genre'])->name('headphones.genre');
 
+});
 
+
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
